# Optimize alphabet cyrillic