--- a/src/alphabets/mod.rs
+++ b/src/alphabets/mod.rs
@@ -1,6 +1,6 @@
-mod cyrillic;
+pub(crate) mod cyrillic;
 pub(crate) mod detection;
-mod latin;
+pub(crate) mod latin;
 
 pub use detection::{detect, raw_detect};