--- a/src/alphabets/mod.rs
+++ b/src/alphabets/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod common;
 pub(crate) mod cyrillic;
 pub(crate) mod detection;
 pub(crate) mod latin;