… src/scripts/detect.rs