Before

```
test bench_alphabet_cyrillic_calculate_scores ... bench:      91,479 ns/iter (+/- 6,833)
test bench_alphabet_latin_calculate_scores    ... bench:       2,459 ns/iter (+/- 204)
test bench_detect                             ... bench:   4,733,183 ns/iter (+/- 269,322)
test bench_detect_script                      ... bench:     110,753 ns/iter (+/- 15,023)

```

After

```
test bench_alphabet_cyrillic_calculate_scores ... bench:       2,176 ns/iter (+/- 449)
test bench_alphabet_latin_calculate_scores    ... bench:       2,562 ns/iter (+/- 610)
test bench_detect                             ... bench:   4,429,384 ns/iter (+/- 203,850)
test bench_detect_script                      ... bench:     112,496 ns/iter (+/- 14,456)

```